Changes all io byte counters to ssize_t

+++ sources/RequestHandler.cpp
@@ -47,7 +47,7 @@ void RequestHandler::handleRequest() {
 }
 
 void RequestHandler::readRequest() {
-	int receivedBytes;
+	ssize_t receivedBytes;
 	char buf[BUFFER_SIZE] = {};
 
 	receivedBytes = recv(_client.fd, buf, BUFFER_SIZE, 0);
